Include a new method in the GpsProvider, GpsRepo and GpsRepoImpl so the related code to obtain an accurate location with a Location listener is in the GpsProvider and not in the MainActivity
