+++ app/src/main/java/com/example/flickrgallery/gps/GpsProvider.kt
@@ -4,7 +4,9 @@ import android.annotation.SuppressLint
 import android.content.Context
 import android.location.Location
 import com.example.flickrgallery.model.GpsSnapshot
+import com.google.android.gms.location.LocationCallback
 import com.google.android.gms.location.LocationRequest
+import com.google.android.gms.location.LocationResult
 import com.google.android.gms.location.LocationServices
 import kotlinx.coroutines.suspendCancellableCoroutine
 import java.util.*
@@ -13,9 +15,12 @@ import kotlin.coroutines.resume
 
 class GpsProvider(context: Context) {
 
+    companion object {
+        private const val ACCEPTABLE_MINIMUM_LOCATION_ACCURACY = 10
+        private const val SECONDS_TO_UPDATE_LOCATION = 5 * 1000L
+    }
 
     private val fusedLocationClient = LocationServices.getFusedLocationProviderClient(context)
-
 
     @SuppressLint("MissingPermission")
     suspend fun getActualLocation(): GpsSnapshot = suspendCancellableCoroutine { continuation ->
@@ -44,5 +49,35 @@ class GpsProvider(context: Context) {
             longitude = 0.0,
             dateCaptured = Calendar.getInstance().timeInMillis
         )
+    }
+
+    @SuppressLint("MissingPermission")
+    fun setAccurateLocationListener(onLocationUpdated: (GpsSnapshot) -> Unit) {
+        val locationCallback = getLocationCallback(onLocationUpdated)
+        fusedLocationClient.requestLocationUpdates(
+                getLocationRequest(), locationCallback, null
+        )
+    }
+
+    private fun getLocationRequest(): LocationRequest {
+        return LocationRequest.create().apply {
+            priority = LocationRequest.PRIORITY_HIGH_ACCURACY
+            interval = SECONDS_TO_UPDATE_LOCATION
+        }
+    }
+
+    private fun getLocationCallback(onLocationUpdated: (GpsSnapshot) -> Unit) = object : LocationCallback() {
+        override fun onLocationResult(locationResult: LocationResult) {
+            for (location in locationResult.locations) {
+                if (isLocationAccurateEnough(location)) {
+                    onLocationUpdated(buildGpsSnapshot(location))
+                    fusedLocationClient.removeLocationUpdates(this)
+                }
+            }
+        }
+    }
+
+    private fun isLocationAccurateEnough(location: Location?): Boolean {
+        return location != null && location.accuracy < ACCEPTABLE_MINIMUM_LOCATION_ACCURACY
     }
 }
