+++ app/src/main/java/com/example/flickrgallery/repo/GpsRepo.kt
@@ -5,4 +5,5 @@ import com.example.flickrgallery.model.GpsSnapshot
 
 interface GpsRepo {
     suspend fun getActualPosition(): GpsSnapshot
+    fun setAccurateLocationListener(onLocationUpdated: (GpsSnapshot) -> Unit)
 }
